--- a/python/packages/sdk/sls_sdk/lib/instrumentation/import_hook.py
+++ b/python/packages/sdk/sls_sdk/lib/instrumentation/import_hook.py
@@ -30,6 +30,17 @@ class ImportHook:
 
     def enable(self, hook_fn):
         self.enabled = True
+
+        already_enabled = bool(
+            [
+                p
+                for p in sys.meta_path
+                if isinstance(p, CustomImporter) and p._module_name == self._module_name
+            ]
+        )
+        if already_enabled:
+            return
+
         self._importer = CustomImporter(self._module_name, hook_fn)
         sys.meta_path.insert(0, self._importer)
         sys.path_importer_cache.clear()
@@ -43,6 +54,12 @@ class ImportHook:
         if undo_hook_fn and self._module_name in sys.modules:
             undo_hook_fn(sys.modules[self._module_name])
 
-        if self._importer and self._importer in sys.meta_path:
-            sys.meta_path.remove(self._importer)
+        importers = [
+            p
+            for p in sys.meta_path
+            if isinstance(p, CustomImporter) and p._module_name == self._module_name
+        ]
+        if importers:
+            for importer in importers:
+                sys.meta_path.remove(importer)
             sys.path_importer_cache.clear()