### Description
* Related issue https://linear.app/serverless/issue/SC-593/python-sdk-big-latency-overhead#comment-448411f1
* Instrument http on the go, without importing the target modules first

### Testing done
Unit, integration & performance tested

#### Integration tests
```
    ✔ success-v3-8 (45192ms)
    ✔ success-v3-9 (8878ms)
    ✔ success-sampled (9453ms)
    ✔ error-v3-8
    ✔ error-v3-9
    ✔ error_unhandled-v3-8
    ✔ error_unhandled-v3-9
    ✔ api_endpoint-rest-api
    ✔ api_endpoint-http-api-v1
    ✔ api_endpoint-http-api-v2
    ✔ api_endpoint-function-url
    ✔ sdk-v3-8
    ✔ sdk-v3-9
    ✔ sdk-dev-mode (3462ms)
    ✔ dashboard-v3-8
    ✔ dashboard-v3-9
    ✔ http_requester-http
    ✔ http_requester-https
    ✔ aiohttp_requester-http
    ✔ aiohttp_requester-https
    ✔ flask_app (421ms)
    ✔ aws_sdk-internal
    ✔ aws_sdk-external (15594ms)
```

#### Performance tests
```
➜  node git:(fix-performance-regression) ✗ npx mocha test/python/aws-lambda-sdk/benchmark/performance.test.js
2023-04-19T13:08:08.749Z ℹ test test uid: omar


  performance
2023-04-19T13:08:08.909Z ℹ test Creating core resources test-python-sdk-omar
2023-04-19T13:08:51.621Z ℹ test Process function success-internal-1
2023-04-19T13:08:51.622Z ℹ test Process function success-internal-2
2023-04-19T13:08:51.623Z ℹ test Process function success-internal-3
2023-04-19T13:08:51.623Z ℹ test Process function success-internal-4
2023-04-19T13:08:51.623Z ℹ test Process function success-internal-5
2023-04-19T13:09:23.033Z ℹ test Cleanup test-python-sdk-omar
2023-04-19T13:09:23.393Z ℹ test Deleted 54 version of the layer test-python-sdk-omar-external
2023-04-19T13:09:23.408Z ℹ test Deleted 54 version of the layer test-python-sdk-omar-internal
2023-04-19T13:09:24.279Z ℹ test Detached IAM policy arn:aws:iam::692593327170:policy/test-python-sdk-omar from test-python-sdk-omar
2023-04-19T13:09:24.512Z ℹ test Deleted IAM role test-python-sdk-omar
2023-04-19T13:09:24.948Z ℹ test Deleted IAM policy arn:aws:iam::692593327170:policy/test-python-sdk-omar
    ✔ should introduce reasonable initialization overhead
    ✔ should introduce reasonable first invocation overhead
    ✔ should introduce reasonable following invocation overhead


  3 passing (1m)
```